Use a single player instance for either audio/video to fix video lag in iOS
